TST: move tests building sdists from test_metadata to test_sdist
